Put exceptions to one place
